==> notebooks/00-metadata_creation.ipynb <==
# Data Catalog Processing

In this notebook I have preserved the steps used to create and preprocess the dataset metadata found in catalog.csv, train.csv, and test.csv.

# libraries

import pandas as pd
import os

data_path = "../data"#PATH TO DATA FOLDER

train_catalog = "../data/public_samples/train.csv"#PATH TO TRAIN.CSV
test_catalog = "../data/public_samples/test.csv"#PATH TO TEXT.CSV

## Adding File Path for .WAV files

All of the original data was in .mp3 format, but we converted those to .wav files. This section adds the path for the .WAV files to the metadata catalogs. The converted WAV files can be found on [Google Drive](https://drive.google.com/drive/folders/1LtRm7rT9IzXIOpSvJQw7GReqS1Wd-j_I?usp=sharing).

train_metadata = pd.read_csv(train_catalog)
test_metadata = pd.read_csv(test_catalog)

# add cols for wav data

# Replace ".mp3" with ".wav" in the "Filename" column
train_metadata['WAV_filename'] = train_metadata['File_Name'].str.replace('.mp3', '.wav')

# Create "WAV_folder" column by concatenating "_WAV" to the "folder" column
train_metadata['WAV_folder'] = train_metadata['folder'] + "_WAV"

# add cols for wav data

# Replace ".mp3" with ".wav" in the "Filename" column
test_metadata['WAV_filename'] = test_metadata['File_Name'].str.replace('.mp3', '.wav')

# Create "WAV_folder" column by concatenating "_WAV" to the "folder" column
test_metadata['WAV_folder'] = test_metadata['folder'] + "_WAV"

## Saving the dataframe

train_metadata.to_csv(train_catalog, index=False)
test_metadata.to_csv(test_catalog, index=False)
